Split mappings main file
Threshold Subgraph recieves events from several sources (contracts). So
a logical division of events handlers can be each contract.

To achieve this, the generic "mappings.ts" file have been splitted in
"pre-app.ts" and "staking.ts".

Also, TokenStaking entities list in subgraph manifest has been updated.
