+++ src/mapping.ts
@@ -7,10 +7,6 @@ import {
   Address,
 } from "@graphprotocol/graph-ts"
 import {
-  OperatorBonded,
-  OperatorConfirmed,
-} from "../generated/SimplePREApplication/SimplePREApplication"
-import {
   Staked,
   ToppedUp,
   Unstaked,
@@ -284,33 +280,6 @@ export function handleDelegateVotesChanged(event: DelegateVotesChanged): void {
   const difference = event.params.newBalance.minus(event.params.previousBalance)
   daoMetric.stakedTotal = daoMetric.stakedTotal.plus(difference)
   daoMetric.save()
-}
-
-export function handleOperatorBonded(event: OperatorBonded): void {
-  const stakingProvider = event.params.stakingProvider
-  const operator = event.params.operator
-  const timestamp = event.params.startTimestamp
-
-  const preApplication = getOrCreatePreApplication(stakingProvider)
-  if (operator === Address.zero()) {
-    store.remove("SimplePREApplication", stakingProvider.toHexString())
-  } else {
-    preApplication.operator = operator
-    preApplication.stake = stakingProvider.toHexString()
-    preApplication.bondedTimestamp = timestamp.plus(BigInt.fromString("12"))
-    preApplication.save()
-  }
-}
-
-export function handleOperatorConfirmed(event: OperatorConfirmed): void {
-  const stakingProvider = event.params.stakingProvider
-  const operator = event.params.operator
-  const timestamp = event.block.timestamp
-
-  const preApplication = getOrCreatePreApplication(stakingProvider)
-  preApplication.operator = operator
-  preApplication.confirmedTimestamp = timestamp
-  preApplication.save()
 }
 
 export function handleMinStakeAmountChanged(
